ap51-flash: Match mac addresses with mask
Different routers will use multiple mac addresses per device. Some of them
use different ones for each port and some of them use one mac for u-boot
and a different one for Linux (depending on the port). The latter makes
problems when detecting the successful flashing of the device. But the
former can result in accidental reflashs of the device when the ethernet
port is switched.

These can be handled by detecting all mac addresses of a device. Each
router_type is now using a mask to limit the number of mac addresses for
each device.

TODO: drop conflicting entries from the list when assigning router_type.

Signed-off-by: Sven Eckelmann <sven.eckelmann@openmesh.com>